requirements.txt
playwright
pandas
sqlalchemy
pydantic
fake-useragent
langchain
langchain-openai
python-dotenv
pymongo
psycopg[binary]

pyyaml
numpy